relaties medicationInventory, medication

+++ src/main/java/nl/wiegersma/dairyfarm/services/MedicationService.java
@@ -3,8 +3,11 @@ package nl.wiegersma.dairyfarm.services;
 import nl.wiegersma.dairyfarm.dtos.MedicationRequestDto;
 import nl.wiegersma.dairyfarm.dtos.MedicationResponseDto;
 import nl.wiegersma.dairyfarm.exceptions.RecordNotFoundException;
+import nl.wiegersma.dairyfarm.exceptions.ResourceInUseException;
 import nl.wiegersma.dairyfarm.mappers.MedicationMapper;
 import nl.wiegersma.dairyfarm.models.Medication;
+import nl.wiegersma.dairyfarm.models.MedicationInventory;
+import nl.wiegersma.dairyfarm.repositories.MedicationInventoryRepository;
 import nl.wiegersma.dairyfarm.repositories.MedicationRepository;
 import org.springframework.stereotype.Service;
 
@@ -15,11 +18,12 @@ public class MedicationService {
 
     private final MedicationMapper medicationMapper;
     private final MedicationRepository medicationRepository;
+    private final MedicationInventoryRepository medicationInventoryRepository;
 
-
-    public MedicationService(MedicationMapper medicationMapper, MedicationRepository medicationRepository) {
+    public MedicationService(MedicationMapper medicationMapper, MedicationRepository medicationRepository, MedicationInventoryRepository medicationInventoryRepository) {
         this.medicationMapper = medicationMapper;
         this.medicationRepository = medicationRepository;
+        this.medicationInventoryRepository = medicationInventoryRepository;
     }
 
     public MedicationResponseDto getMedication(Long id){
@@ -34,8 +38,8 @@ public class MedicationService {
 
     public MedicationResponseDto createMedication(MedicationRequestDto medicationRequestDto){
         Medication medication = medicationMapper.toEntity(medicationRequestDto);
-        medicationRepository.save(medication);
-        return medicationMapper.toDto(medication);
+        Medication updated = medicationRepository.save(medication);
+        return medicationMapper.toDto(updated);
     }
 
     public MedicationResponseDto updateMedication(Long id, MedicationRequestDto medicationRequestDto){
@@ -44,9 +48,13 @@ public class MedicationService {
         return medicationMapper.toDto(updated);
     }
 
-    public void deleteMedication(Long id){
-        Medication medication = medicationRepository.findById(id).orElseThrow(() -> new RecordNotFoundException("medication not found with id: " + id));
-        medicationRepository.delete(medication);
+    public void deleteMedication(Long id) {
+        Medication medication = medicationRepository.findById(id).orElseThrow(() -> new RecordNotFoundException("cant find mediction with id " + id));
+            if (!medication.getMedicationInventories().isEmpty()) {
+                throw new ResourceInUseException("medication is still in use in the medicationInventories");
+            } else {
+                medicationRepository.delete(medication);
+            }
+        }
     }
-}
